requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
